+++ CHANGELOG.md
@@ -11,12 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [hooks:init] Allow to customize help.
+- [child-processes] `DEBUG_CP` environment variable for debugging child processes.
 
 ### Changed
 - Update release notes message.
 - Improve SSE errors logs.
-
-## [2.98.1-beta] - 2020-04-24
 
 ## [2.98.0] - 2020-04-22
 ### Added