+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update release notes message.
 - Improve SSE errors logs.
 - [vtex setup] Add GraphQL files to `lint-staged` config.
+- [vtex workspace] Auto-switch to edition-store when creating new workspace
 
 ## [2.98.0] - 2020-04-22
 ### Added
@@ -47,7 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [clients:sponsor, vtex support] Use HTTPS and new URL `app.io.vtex.com` format.
 - [telemetry:collector] Flush just before exiting process.
 - [sse] Check if token is valid before creating connection.
-- [sse] Abort process on 401 or 403 server errors. 
+- [sse] Abort process on 401 or 403 server errors.
 - [sse:telemetry] Register sse errors with errorKind `SSEError`.
 - Update release notes message.
 
@@ -67,7 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - Command `vtex test e2e` for end-to-end testing.
- 
+
 ## [2.94.0] - 2020-03-31
 ### Added
 - [vtex init] Show `checkout-ui-settings` template for users outside VTEX.